📝 Add a shorter domain
Due to the usage increase of this project by many of you (thanks for the love) I'm adding a shorter custom domain that redirects to the old one, so both of those will work but the docs will look cleaner from now on if you decide to use my domain. Remember you can still deploy it on your Google Cloud account basically for free and you'll have control over the domain.